Move format-console.js under src/util/
