+++ src/classes/TestResult.js
@@ -1,6 +1,6 @@
 import Test from "./Test.js";
 import BubblingEventTarget from "./BubblingEventTarget.js";
-import { stripFormatting } from "../format-console.js";
+import { stripFormatting } from "../util/format-console.js";
 import { delay, formatDuration, interceptConsole, pluralize, stringify } from "../util.js";
 import { formatDiff } from "../util/format-diff.js";
